refactor: dashboard-grid-component
