+++ vue-grid-layout.d.ts
@@ -1,0 +1,7 @@
+declare module "vue-grid-layout" {
+  import Vue from "vue";
+
+  export class GridLayout extends Vue {}
+
+  export class GridItem extends Vue {}
+}
